+++ src/components/PersonalProfile/BasicSection/style.module.scss
@@ -33,7 +33,6 @@
     .school-info {
       display: flex;
       flex-direction: column;
-      justify-content: space-between;
       font-size: 16px;
       > span {
         margin: 2.5px 0;
