+++ src/constants/personalProfile.ts
@@ -1,0 +1,16 @@
+export const basicInfoInit = {
+  img: '',
+  nickName: '',
+  gender: '',
+  college: '',
+  department: '',
+  studentNum: '',
+};
+
+export const moreInfoInit = {
+  birthDate: '',
+  height: '',
+  weight: '',
+  mbti: '',
+  drink: 0,
+};
